Add login/guest flow and modular socket handler
