+++ public/game.html
@@ -41,53 +41,7 @@
       </div>
       <canvas></canvas>
       <!-- username input -->
-      <div
-        style="
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          position: absolute;
-          top: 0;
-          right: 0;
-          bottom: 0;
-          left: 0;
-        "
-      >
-        <form id="usernameForm">
-          <div>
-            <input
-              id="usernameInput"
-              type="text"
-              placeholder="Username"
-              style="
-                padding: 12px;
-                border-radius: 20px;
-                border: none;
-                background: #475569;
-                color: #cbd5e1;
-              "
-            />
-          </div>
-          <div style="text-align: center; margin-top: 8px">
-            <button
-              style="
-                padding: 8px 20px;
-                border-radius: 100px;
-                border: none;
-                background-image: linear-gradient(
-                  to right,
-                  rgb(6, 182, 212),
-                  rgb(59, 130, 246)
-                );
-                color: white;
-                cursor: pointer;
-              "
-            >
-              Start!
-            </button>
-          </div>
-        </form>
-      </div>
+      <div></div>
     </div>
     <!-- username input end -->
     <script src="https://cdn.socket.io/4.6.1/socket.io.min.js"></script>
