+++ readme.md
@@ -9,7 +9,7 @@ npm install
 npm start
 ```
 
-Visit `http://localhost:3000` to see the lobby. You can create or join games in real time and each game runs in its own Socket.IO room.
+Visit `http://localhost:3000` and log in or choose a guest name. After that you can create or join games in real time and each game runs in its own Socket.IO room.
 
 ## Tests
